refactor: remove chat component
