--- src/components/Chat/Chat.css
@@ -1,161 +1,0 @@
-.chat {
-  margin: 0;
-  padding: 0 0 12px;
-  min-height: 580px;
-  box-sizing: border-box;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 24px;
-}
-
-.chat__messages {
-  padding: 6px 15px 15px;
-  width: 100%;
-  flex-direction: column;
-  gap: 24px;
-  max-height: 496px;
-  overflow: scroll;
-  /* Убираем стандартные полосы прокрутки */
-  scrollbar-width: 0; /* Для Firefox */
-  -ms-overflow-style: 0;
-}
-
-.chat__messages::-webkit-scrollbar {
-  width: 0;
-}
-
-.chat__message {
-  position: relative;
-  margin: 0;
-  padding: 16px 44px;
-  max-width: 483px;
-  box-sizing: border-box;
-  display: flex;
-  align-items: center;
-  position: relative;
-  border-radius: 16px;
-  background: no-repeat;
-  border: 1px solid transparent;
-  transition: 0.3s;
-}
-
-.chat__message:not(.chat__message_sender_poll) {
-  box-shadow: var(--box-shadow-black-08-primary);
-  cursor: pointer;
-  transition: 0.3s;
-}
-
-.chat__message:not(.chat__message_sender_poll):hover {
-  border-color: var(--color-blue-2);
-}
-
-.chat__message_sender_bot {
-  align-self: flex-start;
-  justify-content: flex-end;
-  background-color: var(--color-bg-cards-bot);
-  background-image: url(../../assets/icons/carbon_chat-bot.svg);
-  background-position: left 16px top 16px;
-}
-
-.chat__message_sender_user {
-  align-self: flex-end;
-  justify-content: flex-start;
-  background-color: var(--color-bg-cards-user);
-  background-image: url(../../assets/icons/radix-icons_person.svg);
-  background-position: right 16px top 16px;
-}
-
-.chat__message_sender_poll {
-  align-self: flex-end;
-}
-
-.chat__message_text {
-  margin: 0;
-  max-width: 395px;
-  font-size: var(--text-size-s, 14px);
-  line-height: var(--text-line-xs, 20px);
-  word-wrap: break-word;
-}
-
-.chat__message-btn {
-  position: absolute;
-  top: 16px;
-  width: 20px;
-  height: 20px;
-  background: url('../../assets/icons/star-icon-inactive.svg');
-  transition: 0.3s;
-}
-
-.chat__message-btn:hover {
-  filter: var(--filter-brand);
-}
-
-.chat__message-btn_sender_bot {
-  right: 16px;
-}
-
-.chat__message-btn_sender_user {
-  left: 16px;
-}
-
-.chat__message-btn_saved {
-  background-image: url(../../assets/icons/star-icon.svg);
-}
-
-.chat__from {
-  margin: 0;
-  padding: 0 15px;
-  display: flex;
-  justify-content: center;
-  align-items: flex-start;
-  gap: 12px;
-}
-
-.chat__input-wrap {
-  padding: 10px 16px;
-  display: flex;
-  gap: 8px;
-  background-color: var(--color-bg-secondary);
-  border: var(--border-primary);
-  border-radius: 16px;
-  transition: 0.3s;
-}
-
-.chat__input-wrap:hover {
-  background-color: var(--color-bg-tertiary);
-  border: var(--border-brand);
-}
-
-.chat__btn {
-  width: 40px;
-  height: 40px;
-  border-radius: 50%;
-  transition: 0.3s;
-}
-
-.chat__btn:hover {
-  background-color: var(--color-bg-cards-user);
-}
-
-.chat__btn:disabled {
-  background-color: var(--color-bg-secondary);
-}
-
-.chat__mic-btn {
-  background: url(../../assets/icons/microphone.svg) var(--color-bg-brand)
-    no-repeat center / 20px;
-}
-
-.chat__mic-btn:hover {
-  background-color: #ebefff;
-}
-
-.chat__mic-btn_active {
-  transform: scale(1.2);
-}
-
-.chat__send-btn {
-  background: url(../../assets/icons/send-btn.svg) var(--color-bg-brand)
-    no-repeat center / 20px;
-}
