+++ src/components/index.js
@@ -1,4 +1,3 @@
-export * from './Chat/Chat';
 export * from './Footer/Footer';
 export * from './Header/Header';
 export * from './Login/Login';
